+++ R/emis_grid.R
@@ -1,6 +1,6 @@
 #' Allocate emissions into a grid
 #'
-#' @description \code{emis_grid} allocates emissions proportionally to each grid
+#' @description \code{\link{emis_grid}} allocates emissions proportionally to each grid
 #'  cell. The process is performed by intersection between geometries and the grid.
 #' It means that requires "sr" according with your location for the projection.
 #' It is assumed that spobj is a spatial*DataFrame or an "sf" with the pollutants