+++ tests/testthat/test-emis_dist.R
@@ -6,9 +6,9 @@ t1 <- emis_dist(gy = po, spobj = net)
 t2 <- emis_dist(gy = po, spobj = net, pro = pc_profile)
 
 test_that("emis_dist works", {
-  expect_equal(as.numeric(sum(t1$gy)), 1000)
+  expect_equal(as.numeric(sum(t1$emission)), 1000)
 })
 
 test_that("emis_dist works", {
-  expect_equal(as.numeric(sum(t2$h1)), 0.02947961)
+  expect_equal(as.numeric(sum(sf::st_set_geometry(t2, NULL))), 1000)
 })